+++ packages/meteor-components/plugins/componentNameHelper/plugin.js
@@ -1,0 +1,5 @@
+Component.onComponentInitialized(function (component, template) {
+  template.helpers({
+    componentName: function () { return component.name; }
+  });
+});